Lab-03/requirements.txt
numpy
requests
pandas
sqlalchemy
psycopg2
google-cloud-storage
google-cloud-bigquery
